+++ app/src/main/java/nondas/pap/petcareapp/data/network/entity/UserNetworkEntity.kt
@@ -8,5 +8,5 @@ data class UserNetworkEntity(
     val email: String,
     val password: String,
     val name: String,
-    val userId: Int = 0
+    val userId: Long = 0
 )
